requirements.txt
Flask==3.0.3
slack_sdk==3.27.1
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.32.3
openai==1.40.3   # optional, for real AI responses
